usecase 분리
